+++ tests/test_main.py
@@ -105,6 +105,9 @@ def test_main():
                 _main('-f', 'C.yml', '-m' 'a=zab', '-m', 'd=jkl', '--dump')
             assert out.getvalue() == 'a = zab\nb = ghi\nc = xyz\nd = jkl\n'
 
+            # Cover settings
+            _main('-f', 'D.yml', '-s', 'silent=false')
+
 #-------------------------------------------------------------------------------
 
 if __name__ == '__main__': # pragma: no cover