+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 * Added ``-p`` and ``-v`` options.
 * Added support for macros and task references in ``if`` and ``ifnot`` keys.
 * Added ``capture`` section and functionality
+* Added ``settings`` section
 
 0.0.2 (2017-10-26)
 ~~~~~~~~~~~~~~~~~~